requirements.txt
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-apimanagement==4.0.1
azure-mgmt-authorization==4.0.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==32.1.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.6.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.0.0
azure-mgmt-resource==23.2.0
azure-mgmt-sql==3.0.1
azure-mgmt-storage==21.2.1
azure-mgmt-web==7.3.1
msal==1.31.0
requests==2.32.3
